Factor out slots handler into a separate class
